+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Support for Android Studio DMG archives
 - Introduce `VerifyPluginProjectConfigurationTask.hasModulePlugin` to exclude modules using `org.jetbrains.intellij.platform.module` subplugin from `plugin.xml` checks.
 
 ## [2.0.0-beta3] - 2024-05-18